[plc_log] every's package call plc_log to init daily tracing log file

+++ plc_client/src/conf.rs
@@ -25,7 +25,7 @@ pub struct Conf {
     pub service_host: String,
 
     #[arg(long, default_value = "50051")]
-    pub service_port: i32,
+    pub service_port: u16,
 }
 
 impl Conf {
